+++ app/src/main/java/com/example/jetweatherapp/data/remote/models/ApiHourlyWeather.kt
@@ -9,7 +9,7 @@ data class ApiHourlyWeather(
     @SerialName("temperature_2m")
     val temperature2m: List<Double>,
     @SerialName("time")
-    val time: List<Int>,
+    val time: List<Long>,
     @SerialName("weather_code")
     val weatherCode: List<Int>
 )
